refactor(tests): integrate cognito metadata test and rename to test_metadata.py
